requirements.txt
# 2.1.2 used deprecated Instagram app version (269.x) causing challenge_required on login
# 2.4.4 uses updated app version — confirmed working
# Requires Python 3.11 — pydantic-core wheels not available for 3.14
instagrapi==2.4.4
Pillow>=10.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
